+++ src/main/java/org/realityforge/replicant/client/transport/SessionContext.java
@@ -1,11 +1,24 @@
 package org.realityforge.replicant.client.transport;
 
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 public final class SessionContext
 {
+  private String _baseURL;
   private ClientSession _session;
   private RequestEntry _request;
+
+  @Nonnull
+  public String getBaseURL()
+  {
+    return _baseURL;
+  }
+
+  public void setBaseURL( @Nonnull final String baseURL )
+  {
+    _baseURL = baseURL;
+  }
 
   @Nullable
   public ClientSession getSession()